Address part of review comments
